Добавил контроллер, который нужен был в первом задании
